ch4/am: refactor mpidig rma protocols
