+++ src/mpid/ch4/generic/mpidig_init.c
@@ -58,104 +58,22 @@ int MPIDIG_init(void)
     mpi_errno = MPIDIG_am_long_init();
     MPIR_ERR_CHECK(mpi_errno);
 
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_PUT_REQ, &MPIDIG_put_origin_cb, &MPIDIG_put_target_msg_cb);
+    mpi_errno = MPIDIG_am_put_init();
     MPIR_ERR_CHECK(mpi_errno);
 
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_PUT_ACK, NULL, &MPIDIG_put_ack_target_msg_cb);
+    mpi_errno = MPIDIG_am_get_init();
     MPIR_ERR_CHECK(mpi_errno);
 
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_GET_REQ, &MPIDIG_get_origin_cb, &MPIDIG_get_target_msg_cb);
+    mpi_errno = MPIDIG_am_cswap_init();
     MPIR_ERR_CHECK(mpi_errno);
 
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_GET_ACK,
-                                 &MPIDIG_get_ack_origin_cb, &MPIDIG_get_ack_target_msg_cb);
+    mpi_errno = MPIDIG_am_acc_init();
     MPIR_ERR_CHECK(mpi_errno);
 
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_CSWAP_REQ,
-                                 &MPIDIG_cswap_origin_cb, &MPIDIG_cswap_target_msg_cb);
+    mpi_errno = MPIDIG_am_get_acc_init();
     MPIR_ERR_CHECK(mpi_errno);
 
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_CSWAP_ACK,
-                                 &MPIDIG_cswap_ack_origin_cb, &MPIDIG_cswap_ack_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_ACC_REQ, &MPIDIG_acc_origin_cb, &MPIDIG_acc_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_GET_ACC_REQ,
-                                 &MPIDIG_get_acc_origin_cb, &MPIDIG_get_acc_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_ACC_ACK, NULL, &MPIDIG_acc_ack_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_GET_ACC_ACK,
-                                 &MPIDIG_get_acc_ack_origin_cb, &MPIDIG_get_acc_ack_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_WIN_COMPLETE, NULL, &MPIDIG_win_ctrl_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_WIN_POST, NULL, &MPIDIG_win_ctrl_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_WIN_LOCK, NULL, &MPIDIG_win_ctrl_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_WIN_LOCK_ACK, NULL, &MPIDIG_win_ctrl_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_WIN_UNLOCK, NULL, &MPIDIG_win_ctrl_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_WIN_UNLOCK_ACK, NULL, &MPIDIG_win_ctrl_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_WIN_LOCKALL, NULL, &MPIDIG_win_ctrl_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_WIN_LOCKALL_ACK, NULL, &MPIDIG_win_ctrl_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_WIN_UNLOCKALL, NULL, &MPIDIG_win_ctrl_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_WIN_UNLOCKALL_ACK, NULL, &MPIDIG_win_ctrl_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_PUT_IOV_REQ,
-                                 &MPIDIG_put_iov_origin_cb, &MPIDIG_put_iov_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_PUT_IOV_ACK, NULL, &MPIDIG_put_iov_ack_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_PUT_DAT_REQ,
-                                 &MPIDIG_put_data_origin_cb, &MPIDIG_put_data_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_ACC_IOV_REQ,
-                                 &MPIDIG_acc_iov_origin_cb, &MPIDIG_acc_iov_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_GET_ACC_IOV_REQ,
-                                 &MPIDIG_get_acc_iov_origin_cb, &MPIDIG_get_acc_iov_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_ACC_IOV_ACK, NULL, &MPIDIG_acc_iov_ack_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_GET_ACC_IOV_ACK,
-                                 NULL, &MPIDIG_get_acc_iov_ack_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_ACC_DAT_REQ,
-                                 &MPIDIG_acc_data_origin_cb, &MPIDIG_acc_data_target_msg_cb);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_am_reg_cb(MPIDIG_GET_ACC_DAT_REQ,
-                                 &MPIDIG_get_acc_data_origin_cb,
-                                 &MPIDIG_get_acc_data_target_msg_cb);
+    mpi_errno = MPIDIG_am_win_sync_init();
     MPIR_ERR_CHECK(mpi_errno);
 
     mpi_errno = MPIDIG_am_comm_abort_init();
@@ -164,9 +82,6 @@ int MPIDIG_init(void)
     MPIDIU_map_create((void **) &(MPIDI_global.win_map), MPL_MEM_RMA);
 
     mpi_errno = MPIDIG_RMA_Init_sync_pvars();
-    MPIR_ERR_CHECK(mpi_errno);
-
-    mpi_errno = MPIDIG_RMA_Init_targetcb_pvars();
     MPIR_ERR_CHECK(mpi_errno);
 
     MPIDI_global.is_ch4u_initialized = 1;
