+++ src/mpid/ch4/include/mpidch4r.h
@@ -18,8 +18,6 @@
 #include "ch4r_probe.h"
 #include "ch4r_proc.h"
 #include "ch4r_request.h"
-#include "ch4r_rma_origin_callbacks.h"
-#include "ch4r_rma_target_callbacks.h"
 #include "ch4r_rma.h"
 #include "ch4r_win.h"
 #include "ch4r_buf.h"
