+++ src/mpid/ch4/src/ch4_types.h
@@ -127,82 +127,6 @@ typedef struct MPIDIG_win_cntrl_msg_t {
     int16_t lock_type;
 } MPIDIG_win_cntrl_msg_t;
 
-typedef struct MPIDIG_put_msg_t {
-    int src_rank;
-    uint64_t win_id;
-    MPIR_Request *preq_ptr;
-    MPI_Aint target_disp;
-    uint64_t count;
-    MPI_Datatype datatype;
-    int n_iov;
-} MPIDIG_put_msg_t;
-
-typedef struct MPIDIG_put_iov_ack_msg_t {
-    int src_rank;
-    MPIR_Request *target_preq_ptr;
-    MPIR_Request *origin_preq_ptr;
-} MPIDIG_put_iov_ack_msg_t;
-typedef MPIDIG_put_iov_ack_msg_t MPIDIG_acc_iov_ack_msg_t;
-typedef MPIDIG_put_iov_ack_msg_t MPIDIG_get_acc_iov_ack_msg_t;
-
-typedef struct MPIDIG_put_dat_msg_t {
-    MPIR_Request *preq_ptr;
-} MPIDIG_put_dat_msg_t;
-typedef MPIDIG_put_dat_msg_t MPIDIG_acc_dat_msg_t;
-typedef MPIDIG_put_dat_msg_t MPIDIG_get_acc_dat_msg_t;
-
-typedef struct MPIDIG_put_ack_msg_t {
-    MPIR_Request *preq_ptr;
-} MPIDIG_put_ack_msg_t;
-
-typedef struct MPIDIG_get_msg_t {
-    int src_rank;
-    uint64_t win_id;
-    MPIR_Request *greq_ptr;
-    MPI_Aint target_disp;
-    uint64_t count;
-    MPI_Datatype datatype;
-    int n_iov;
-} MPIDIG_get_msg_t;
-
-typedef struct MPIDIG_get_ack_msg_t {
-    MPIR_Request *greq_ptr;
-} MPIDIG_get_ack_msg_t;
-
-typedef struct MPIDIG_cswap_req_msg_t {
-    int src_rank;
-    uint64_t win_id;
-    MPIR_Request *req_ptr;
-    MPI_Aint target_disp;
-    MPI_Datatype datatype;
-} MPIDIG_cswap_req_msg_t;
-
-typedef struct MPIDIG_cswap_ack_msg_t {
-    MPIR_Request *req_ptr;
-} MPIDIG_cswap_ack_msg_t;
-
-typedef struct MPIDIG_acc_req_msg_t {
-    int src_rank;
-    uint64_t win_id;
-    MPIR_Request *req_ptr;
-    int origin_count;
-    MPI_Datatype origin_datatype;
-    int target_count;
-    MPI_Datatype target_datatype;
-    MPI_Op op;
-    MPI_Aint target_disp;
-    uint64_t result_data_sz;
-    int n_iov;
-} MPIDIG_acc_req_msg_t;
-
-typedef struct MPIDIG_acc_req_msg_t MPIDIG_get_acc_req_msg_t;
-
-typedef struct MPIDIG_acc_ack_msg_t {
-    MPIR_Request *req_ptr;
-} MPIDIG_acc_ack_msg_t;
-
-typedef MPIDIG_acc_ack_msg_t MPIDIG_get_acc_ack_msg_t;
-
 typedef struct MPIDIG_comm_req_list_t {
     MPIR_Comm *comm[2][4];
     MPIDIG_rreq_t *uelist[2][4];
